integration testing

+++ src/CurrentSensor.h
@@ -35,6 +35,7 @@ When using internal Clock (need CNVST?) are we doing this?
 */
 
 
+
 /* ADC MODE CONTROL */
 
 // bit field - Table 2
@@ -211,7 +212,7 @@ public:
 
   uint16_t read_data(current_sensor_channel_t channel);
   float read_voltage(current_sensor_channel_t channel);
-  float average_voltage_readings(current_sensor_channel_t channel, uint16_t num_samples);
+  float average_voltage_readings(current_sensor_channel_t channel, uint32_t num_samples);
 private:
   uint16_t read_write_data(uint16_t data);
   uint8_t chip_select;
